requirements.txt
asgiref==3.11.1
Django==4.2.28
django-environ==0.12.1
django-filter==25.1
django-jazzmin==3.0.2
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
drf-spectacular==0.29.0
django-cors-headers>=4.0.0
inflection==0.5.1
jsonschema==4.26.0
pillow==12.1.1
PyJWT==2.11.0
PyYAML==6.0.3
sqlparse==0.5.5
tzdata==2025.3
